# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Enterprise OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leaves:
    get:
      operationId: listLeaves
      tags: [leaves]
      summary: List leave requests
      responses:
        "200":
          description: Leave requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaveRequest"
    post:
      operationId: createLeave
      tags: [leaves]
      summary: Create leave request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaveInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"
  /leaves/{id}/status:
    patch:
      operationId: updateLeaveStatus
      tags: [leaves]
      summary: Approve or reject leave
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaveStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"

  # --- LEADS ---
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products/inventory
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: category
          schema:
            type: string
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update product
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects
      parameters:
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: Projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update project
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete project
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # --- TASKS ---
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks
      parameters:
        - in: query
          name: project_id
          schema:
            type: integer
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: assignee_id
          schema:
            type: integer
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # --- MILESTONES ---
  /milestones:
    get:
      operationId: listMilestones
      tags: [milestones]
      summary: List milestones
      parameters:
        - in: query
          name: project_id
          schema:
            type: integer
      responses:
        "200":
          description: Milestones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Milestone"
    post:
      operationId: createMilestone
      tags: [milestones]
      summary: Create milestone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MilestoneInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Milestone"
  /milestones/{id}:
    patch:
      operationId: updateMilestone
      tags: [milestones]
      summary: Update milestone
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MilestoneUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Milestone"

  # --- ANALYTICS ---
...
components:
  schemas:
...
    LeaveRequest:
      type: object
      required: [id, employee_id, employee_name, type, start_date, end_date, status, reason]
      properties:
        id:
          type: integer
        employee_id:
          type: integer
        employee_name:
          type: string
        type:
          type: string
        start_date:
          type: string
        end_date:
          type: string
        status:
          type: string
        reason:
          type: string
    LeaveInput:
      type: object
      required: [employee_id, type, start_date, end_date, reason]
      properties:
        employee_id:
          type: integer
        type:
          type: string
        start_date:
          type: string
        end_date:
          type: string
        reason:
          type: string
    LeaveStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string

    # --- LEADS ---
...
    Product:
      type: object
      required: [id, name, category, sku, stock, unit_price, status]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        sku:
          type: string
        stock:
          type: integer
        unit_price:
          type: number
        status:
          type: string
        description:
          type: ["string", "null"]
        vendor:
          type: ["string", "null"]
    ProductInput:
      type: object
      required: [name, category, sku, stock, unit_price]
      properties:
        name:
          type: string
        category:
          type: string
        sku:
          type: string
        stock:
          type: integer
        unit_price:
          type: number
        status:
          type: string
        description:
          type: string
        vendor:
          type: string
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        stock:
          type: integer
        unit_price:
          type: number
        status:
          type: string
        description:
          type: string
        vendor:
          type: string
...
    Project:
      type: object
      required: [id, name, status, progress, start_date, end_date]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
        progress:
          type: integer
        start_date:
          type: string
        end_date:
          type: string
        manager:
          type: ["string", "null"]
        team:
          type: array
          items:
            type: string
        priority:
          type: ["string", "null"]
    ProjectInput:
      type: object
      required: [name, status, start_date, end_date]
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
        progress:
          type: integer
        start_date:
          type: string
        end_date:
          type: string
        manager:
          type: string
        priority:
          type: string
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
        progress:
          type: integer
        end_date:
          type: string
        manager:
          type: string
        priority:
          type: string

    # --- TASKS ---
    Task:
      type: object
      required: [id, title, status, priority, created_at]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
        priority:
          type: string
        project_id:
          type: ["integer", "null"]
        project_name:
          type: ["string", "null"]
        assignee:
          type: ["string", "null"]
        due_date:
          type: ["string", "null"]
        created_at:
          type: string
    TaskInput:
      type: object
      required: [title, status, priority]
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
        priority:
          type: string
        project_id:
          type: integer
        assignee:
          type: string
        due_date:
          type: string
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
        priority:
          type: string
        assignee:
          type: string
        due_date:
          type: string

    # --- MILESTONES ---
    Milestone:
      type: object
      required: [id, title, project_id, due_date, status]
      properties:
        id:
          type: integer
        title:
          type: string
        project_id:
          type: integer
        project_name:
          type: ["string", "null"]
        due_date:
          type: string
        status:
          type: string
        description:
          type: ["string", "null"]
    MilestoneInput:
      type: object
      required: [title, project_id, due_date]
      properties:
        title:
          type: string
        project_id:
          type: integer
        due_date:
          type: string
        status:
          type: string
        description:
          type: string
    MilestoneUpdate:
      type: object
      properties:
        title:
          type: string
        due_date:
          type: string
        status:
          type: string
        description:
          type: string

    # --- ANALYTICS ---
...